+++ rumydata/cell/rule.py
@@ -369,3 +369,21 @@ def make_static_cell_rule(func, assertion, exception=ex.UrNotMyDataError) -> Rul
             return assertion
 
     return FactoryRule()
+
+
+class ColumnComparisonRule(Rule):
+    exception_class = ex.ColumnComparisonError
+
+    def __init__(self, compare_to: str):
+        self.compare_to = compare_to
+
+    def prepare(self, data: CellData) -> tuple:
+        return data.value, data.compare[self.compare_to]
+
+
+class GreaterThanColumn(ColumnComparisonRule):
+    def evaluator(self):
+        return lambda x, y: x > y
+
+    def explain(self) -> str:
+        return f"must be greater than column '{self.compare_to}'"
